Configuration changes for database and error reporting
PDO objects are used throughout the tests and can't be serialised, so we have to disable backups of globals and static attributes.

Error reporting is based on the PHP defaults, which don't include (depending on version) E_NOTICE, E_STRICT and E_DEPRECATED, but tests should run with all warnings and errors. PHP constants can't be used outside of php.ini so we have to use a large integer value instead for future-proofing.