Change Earned Badge to have integer ID primary key
Although the 'real' primary key is a text UID field, it is easier to keep with the convention of having an id INTEGER PRIMARY KEY column on all schemas. In order to allow lookups based on UID a new function getIdFromUid has been added, along with tests to ensure its functionality is correct.