Move URL fields out of database and generate dynamically instead
By generating the URL on a per-request basis, we are removing a hardcoded variable, making it easier to deploy the application elsewhere, as well as removing the risk of information in the database falling out of sync with the web server configuration.

In order to ensure that the transition was successful, the port of the test server was changed to 8080, as this would cause tests relying on the old data fields to fail.